requirements.txt
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.12.0
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
supervisor==4.2.5
cryptography==41.0.3
schedule==1.2.1
